+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Change Log
 
-## [0.2.0]
+## [0.2.0] 2021-2-6
 
-- Added commands for GDB
+- Introduce a few CMakeTools build kit -specific commands `cmake.buildkit.*.exe` to enable the use of kit-specific CMake & generator tools instead of the generic MSYS2-specific ones.
 
-- Documentation update.
+- Per user settings template update
+
+- Documentation update
 
 ## [0.1.0] 2021-1-30